requirements.txt
google-cloud-logging~=3.5.0
requests~=2.28.1